requirements.txt
# Core Data Science and ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning
torch>=1.10.0
transformers>=4.15.0
tensorflow>=2.8.0

# NLP and Text Processing
spacy>=3.2.0
nltk>=3.6.0
textblob>=0.17.0
gensim>=4.0.0

# Vector Databases and Retrieval
faiss-cpu>=1.7.0
pinecone-client>=2.2.0
langchain>=0.0.200

# REST API
fastapi>=0.79.0
uvicorn>=0.18.0
pydantic>=1.9.0

# MLOps and Experiment Tracking
mlflow>=1.26.0
wandb>=0.13.0

# Deployment and Configuration
python-dotenv>=0.19.0
pyyaml>=5.4.0

# Testing and Development
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7b0
flake8>=3.9.0
mypy>=0.910

# Utilities
tqdm>=4.60.0
requests>=2.26.0
python-multipart>=0.0.5
